requirements.txt
absl-py==1.3.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
appdirs==1.4.4
#apturl==0.5.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
audioread==2.1.9
Automat==0.8.0
backcall==0.2.0
bayesian-optimization==1.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blessed==1.17.10
#blinker==1.4
#Brlapi==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
clang==5.0
Click==7.0
#cloud-init==22.2
cloudpickle==1.6.0
colorama==0.4.3
colorful==0.5.4
command-not-found==0.3
configobj==5.0.6
configparser==5.1.0
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
dill==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.5
dnspython==1.16.0
docker-pycreds==0.4.0
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
filelock==3.0.12
flatbuffers==22.10.26
fsspec==2022.1.0
future==0.18.2
gast==0.4.0
gin-config==0.3.0
gitdb==4.0.9
GitPython==3.1.24
google==3.0.0
google-api-core==1.23.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpg===1.13.1-unknown
gpustat==1.0.0.dev0
grpcio==1.41.1
gym==0.17.3
h5py==3.1.0
hiredis==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
imgaug==0.4.0
importlib-metadata==1.5.0
importlib-resources==3.0.0
imutils==0.5.4
incremental==16.10.1
iotop==0.6
ipython==8.2.0
jedi==0.18.1
Jinja2==2.10.1
joblib==0.17.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kaggle==1.5.12
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tcn==3.4.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
librosa==0.8.0
llvmlite==0.35.0
lmdb==1.3.0
lockfile==0.12.2
louis==3.12.0
lz4==3.1.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.3.2
matplotlib-inline==0.1.3
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
multidict==5.1.0
netifaces==0.10.4
networkx==2.5
numba==0.52.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.11
opencensus-context==0.1.2
opencv-python==4.4.0.46
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.4
paramiko==2.6.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pooch==1.3.0
progress==1.6
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.1
psutil==5.7.2
pudb==2020.1
pure-eval==0.2.2
py-spy==0.3.3
pyaml==20.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyDeprecate==0.3.1
pygame==2.0.0.dev10
pyglet==1.5.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-slugify==5.0.2
pytorch-lightning==1.5.8
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
ray==1.9.2
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-optimize==0.8.1
scipy==1.9.3
seaborn==0.11.0
SecretStorage==2.3.1
sentry-sdk==1.5.0
service-identity==18.1.0
setproctitle==1.3.2
Shapely==1.7.1
shortuuid==1.0.8
simplejson==3.16.0
six==1.15.0
smmap==5.0.0
sos==4.3
SoundFile==0.10.3.post1
soupsieve==2.0.1
ssh-import-id==5.10
stack-data==0.2.0
subprocess32==3.5.4
systemd-python==234
tabulate==0.8.7
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.10.0
tensorflow-addons==0.11.2
tensorflow-datasets==4.4.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==0.24.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.11.26
timm==0.5.4
torch==1.13.0
torchaudio==0.13.0
torchinfo==1.6.3
torchmetrics==0.7.0
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.50.2
traitlets==5.1.1
Twisted==18.9.0
typeguard==2.10.0
typing_extensions==4.0.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
urwid==2.1.2
usb-creator==0.3.7
wadllib==1.3.3
wandb==0.13.4
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
xkit==0.0.0
yarl==1.6.3
yaspin==2.1.0
zipp==1.0.0
zope.interface==4.7.1